+++ lib/src/features/show spaces/all space mvvm/all_spaces_vm.dart
@@ -1,9 +1,7 @@
-import 'dart:developer';
 import 'package:git_flutter_festou/src/core/exceptions/repository_exception.dart';
 import 'package:git_flutter_festou/src/core/fp/either.dart';
 import 'package:git_flutter_festou/src/core/providers/application_providers.dart';
-import 'package:git_flutter_festou/src/features/home/all%20space%20mvvm/all_spaces_state.dart';
-import 'package:git_flutter_festou/src/models/space_with_image_model.dart';
+import 'package:git_flutter_festou/src/features/show%20spaces/all%20space%20mvvm/all_spaces_state.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'all_spaces_vm.g.dart';
@@ -20,12 +18,8 @@ class AllSpacesVm extends _$AllSpacesVm {
 
       switch (spacesResult) {
         case Success(value: final spacesData):
-          log('$spacesData');
-
-          final spaces = <SpaceWithImages>[];
-          spaces.addAll(spacesData);
           return AllSpaceState(
-              status: AllSpaceStateStatus.loaded, spaces: spaces);
+              status: AllSpaceStateStatus.loaded, spaces: spacesData);
 
         case Failure(exception: RepositoryException(:final message)):
           errorMessage = message;
@@ -35,7 +29,7 @@ class AllSpacesVm extends _$AllSpacesVm {
           );
       }
     } on Exception {
-      errorMessage = 'Erro desconhecido'; // Atualize a mensagem de erro
+      errorMessage = 'Erro desconhecido';
 
       return AllSpaceState(status: AllSpaceStateStatus.loaded, spaces: []);
     }
