+++ lib/src/features/show spaces/my favorite spaces mvvm/my_favorite_spaces_page.dart
@@ -2,9 +2,10 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:git_flutter_festou/src/core/ui/helpers/messages.dart';
-import 'package:git_flutter_festou/src/features/home/my%20favorite%20spaces%20mvvm/my_favorite_spaces_state.dart';
-import 'package:git_flutter_festou/src/features/home/my%20favorite%20spaces%20mvvm/my_favorite_spaces_vm.dart';
-import 'package:git_flutter_festou/src/features/home/widgets/space_card.dart';
+import 'package:git_flutter_festou/src/features/show%20spaces/widgets/my_sliver_list.dart';
+import 'package:git_flutter_festou/src/features/show%20spaces/widgets/my_sliver_to_box_adapter.dart';
+import 'package:git_flutter_festou/src/features/show%20spaces/my%20favorite%20spaces%20mvvm/my_favorite_spaces_state.dart';
+import 'package:git_flutter_festou/src/features/show%20spaces/my%20favorite%20spaces%20mvvm/my_favorite_spaces_vm.dart';
 
 class MyFavoriteSpacePage extends ConsumerStatefulWidget {
   const MyFavoriteSpacePage({super.key});
@@ -35,24 +36,10 @@ class _MyFavoriteSpacePageState extends ConsumerState<MyFavoriteSpacePage> {
         data: (MyFavoriteSpacesState data) {
           return CustomScrollView(
             slivers: [
-              //paras colocar widgets nao rolaveis.
-              const SliverToBoxAdapter(
-                child: Center(
-                    child: Text(
-                  'MY FAVORITE SPACES',
-                  style: TextStyle(
-                    fontSize: 28,
-                    fontWeight: FontWeight.bold,
-                  ),
-                )),
+              const MySliverToBoxAdapter(
+                text: 'MY FAVORITE SPACES',
               ),
-              SliverList(
-                delegate: SliverChildBuilderDelegate(
-                    (context, index) => Column(
-                          children: [SpaceCard2(space: data.spaces[index])],
-                        ),
-                    childCount: data.spaces.length),
-              ),
+              MySliverList(data: data, spaces: favSpaces),
             ],
           );
         },
