+++ lib/src/features/show spaces/my favorite spaces mvvm/my_favorite_spaces_state.dart
