+++ lib/src/features/show spaces/my favorite spaces mvvm/my_favorite_spaces_vm.dart
@@ -1,7 +1,7 @@
 import 'package:git_flutter_festou/src/core/exceptions/repository_exception.dart';
 import 'package:git_flutter_festou/src/core/fp/either.dart';
 import 'package:git_flutter_festou/src/core/providers/application_providers.dart';
-import 'package:git_flutter_festou/src/features/home/my%20favorite%20spaces%20mvvm/my_favorite_spaces_state.dart';
+import 'package:git_flutter_festou/src/features/show%20spaces/my%20favorite%20spaces%20mvvm/my_favorite_spaces_state.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'my_favorite_spaces_vm.g.dart';
@@ -27,7 +27,7 @@ class MyFavoriteSpacesVm extends _$MyFavoriteSpacesVm {
               status: MyFavoriteSpacesStateStatus.error, spaces: []);
       }
     } on Exception {
-      errorMessage = 'Erro desconhecido'; // Atualize a mensagem de erro
+      errorMessage = 'Erro desconhecido';
 
       return MyFavoriteSpacesState(
           status: MyFavoriteSpacesStateStatus.loaded, spaces: []);
