+++ lib/src/features/show spaces/my space mvvm/my_spaces_vm.dart
@@ -1,8 +1,7 @@
 import 'package:git_flutter_festou/src/core/exceptions/repository_exception.dart';
 import 'package:git_flutter_festou/src/core/fp/either.dart';
 import 'package:git_flutter_festou/src/core/providers/application_providers.dart';
-import 'package:git_flutter_festou/src/features/home/my%20space%20mvvm/my_spaces_state.dart';
-import 'package:git_flutter_festou/src/models/space_with_image_model.dart';
+import 'package:git_flutter_festou/src/features/show%20spaces/my%20space%20mvvm/my_spaces_state.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'my_spaces_vm.g.dart';
@@ -13,24 +12,20 @@ class MySpacesVm extends _$MySpacesVm {
   @override
   Future<MySpacesState> build() async {
     final spaceRepository = ref.read(spaceFirestoreRepositoryProvider);
-    //final BarbershopModel(id: barbershopId) =
-    //await ref.read(getMyBarbershopProvider.future);
     try {
       final spacesResult = await spaceRepository.getMySpaces();
 
       switch (spacesResult) {
         case Success(value: final spacesData):
-          final spaces = <SpaceWithImages>[];
-          spaces.addAll(spacesData);
           return MySpacesState(
-              status: MySpacesStateStatus.loaded, spaces: spaces);
+              status: MySpacesStateStatus.loaded, spaces: spacesData);
 
         case Failure(exception: RepositoryException(:final message)):
           errorMessage = message;
           return MySpacesState(status: MySpacesStateStatus.error, spaces: []);
       }
     } on Exception {
-      errorMessage = 'Erro desconhecido'; // Atualize a mensagem de erro
+      errorMessage = 'Erro desconhecido';
 
       return MySpacesState(status: MySpacesStateStatus.loaded, spaces: []);
     }
